Moved the preshooter into the flywheel subsystem.
The preshooter is moved into the flywheel subsystem, it is renamed the feeder, indexerproto has been renamed to indexer.
