+++ src/main/java/frc/robot/commands/TestIndexBelt.java
@@ -1,13 +1,13 @@
 package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.IndexerProto;
+import frc.robot.subsystems.Indexer;
 
 public class TestIndexBelt extends CommandBase{
 
     @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-    private IndexerProto sys_indexer;
+    private Indexer sys_indexer;
 
-    public TestIndexBelt(IndexerProto subsystem){
+    public TestIndexBelt(Indexer subsystem){
         sys_indexer = subsystem; 
 
         addRequirements(subsystem);
