--- src/main/java/frc/robot/commands/TuningTesting.java
@@ -1,48 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Colour;
-
-public class TuningTesting extends CommandBase {
-
-  private Colour sys_colour;
-
-  /**
-   * Creates a new TuningTesting.
-   */
-  public TuningTesting(Colour subsystem) {
-    // Use addRequirements() here to declare subsystem dependencies.
-
-    sys_colour = subsystem;
-    addRequirements(sys_colour);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    sys_colour.initialColourCalibration();
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return true;
-  }
-}
