+++ src/main/java/frc/robot/commands/training/TrainerLookShooterState.java
@@ -5,10 +5,8 @@ import org.jetbrains.annotations.NotNull;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.robot.Constants;
 import frc.robot.base.StateCommandBase;
-import frc.robot.subsystems.IndexerProto;
 import frc.robot.subsystems.Limelight;
 import frc.robot.subsystems.Limelight.TargetType;
-import frc.robot.subsystems.shooter.ShooterFlywheel;
 import frc.robot.subsystems.shooter.ShooterTurret;
 import frc.robot.training.TrainerDashboard;
 import frc.robot.training.Setpoint;
